Refactor layout and add new components for wallet connection and cache status.
- Replaced Providers with RainbowKitProvider in layout.tsx.
- Introduced BlockchainSelector, CacheStatusDashboard, and ConnectWallet components.
- Removed old ConnectWallet implementation and added a new version without Providers.
- Created Header component to include navigation links and wallet connection.
