+++ src/components/CacheStatusDashboard.tsx
