+++ src/components/ConnectWallet.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import { ConnectButton } from '@rainbow-me/rainbowkit';
+
+export default function ConnectWallet() {
+  return (
+    <div className=''>
+      <ConnectButton
+        label='Connect Wallet'
+        accountStatus={'full'}
+        chainStatus={'name'}
+      />
+    </div>
+  );
+}
