+++ src/components/Header.tsx
@@ -2,8 +2,8 @@ import Link from 'next/link';
 import Image from 'next/image';
 import logo from 'public/logo.svg';
 import horn from 'public/icons/horn.svg';
-// import BlockchainSelector from './blockchainSelector';
-import ConnectWallet from './connectWallet';
+// import BlockchainSelector from './BlockchainSelector';
+import ConnectWallet from './ConnectWallet';
 export default function Header() {
   return (
     <div className='w-full bg-black text-white z-10 flex flex-wrap items-center justify-between p-4 px-10'>
